v4/train/requirements.txt
mujoco>=3.0.0
stable-baselines3>=2.2.1
gymnasium>=0.29.1
flask>=3.0.0
flask-socketio>=5.3.6
numpy>=1.26.0
torch>=2.1.0
eventlet>=0.34.0
tensorboard>=2.15.0
tqdm>=4.66.0
rich>=13.0.0
Pillow>=10.0.0
